[ DONE ] 1. Добавил метод getFieldMessage для ValidateResult; 2.  Сделал поле ACL публичным для Service; 3. Добавил проверку доступа для репозитория;

+++ src/front/Repository.ts
@@ -83,21 +83,23 @@ export class Repository<
   public async init(): Promise<void> {
     if (!this.isInit) {
       try {
-        this.startLoad();
+        if (this.service.ACL.collection.includes(this.service.group)) {
+          this.startLoad();
 
-        const collection: T[] | void = await this.service.collection();
+          const collection: T[] | void = await this.service.collection();
 
-        if (isArray(collection)) {
-          runInAction(`[ ${this.constructor.name} ][ INIT ][ SUCCESS ]`, () => {
-            this.collection = collection.reduce<ObservableMap<string, T>>(
-              (preValue, item: T) => preValue.set(item.id, item),
-              observable.map(),
-            );
+          if (isArray(collection)) {
+            runInAction(`[ ${this.constructor.name} ][ INIT ][ SUCCESS ]`, () => {
+              this.collection = collection.reduce<ObservableMap<string, T>>(
+                (preValue, item: T) => preValue.set(item.id, item),
+                observable.map(),
+              );
 
-            this.isInit = true;
-          });
-        } else {
-          throw new Error(`collection: ${Object.prototype.toString.call(collection)}`);
+              this.isInit = true;
+            });
+          } else {
+            throw new Error(`collection: ${Object.prototype.toString.call(collection)}`);
+          }
         }
       } catch (error) {
         console.error(`[ ${this.constructor.name} ][ INIT ][ ERROR_MESSAGE: ${error.message || error} ]`);
@@ -112,19 +114,21 @@ export class Repository<
   public async create(data: object): Promise<T | void> {
     if (this.isInit) {
       try {
-        this.startLoad();
+        if (this.service.ACL.create.includes(this.service.group)) {
+          this.startLoad();
 
-        const item: T | void = await this.service.create(data);
+          const item: T | void = await this.service.create(data);
 
-        if (item) {
-          runInAction(`[ ${this.constructor.name} ][ CREATE ][ SUCCESS ]`, () => {
-            this.collection.set(item.id, item);
-          });
-        } else {
-          throw new Error(`item: ${Object.prototype.toString.call(item)}`);
+          if (item) {
+            runInAction(`[ ${this.constructor.name} ][ CREATE ][ SUCCESS ]`, () => {
+              this.collection.set(item.id, item);
+            });
+          } else {
+            throw new Error(`item: ${Object.prototype.toString.call(item)}`);
+          }
+
+          return item;
         }
-
-        return item;
       } catch (error) {
         console.error(`[ ${this.constructor.name} ][ CREATE ][ ERROR_MESSAGE: ${error.message || error} ]`);
 
@@ -142,19 +146,21 @@ export class Repository<
   public async update(data: object): Promise<T | void> {
     if (this.isInit) {
       try {
-        this.startLoad();
+        if (this.service.ACL.update.includes(this.service.group)) {
+          this.startLoad();
 
-        const item: T | void = await this.service.update(data);
+          const item: T | void = await this.service.update(data);
 
-        if (item) {
-          runInAction(`[ ${this.constructor.name} ][ UPDATE ][ SUCCESS ]`, () => {
-            this.collection.set(item.id, item);
-          });
-        } else {
-          throw new Error(`item: ${Object.prototype.toString.call(item)}`);
+          if (item) {
+            runInAction(`[ ${this.constructor.name} ][ UPDATE ][ SUCCESS ]`, () => {
+              this.collection.set(item.id, item);
+            });
+          } else {
+            throw new Error(`item: ${Object.prototype.toString.call(item)}`);
+          }
+
+          return item;
         }
-
-        return item;
       } catch (error) {
         console.error(`[ ${this.constructor.name} ][ UPDATE ][ ERROR_MESSAGE: ${error.message || error} ]`);
 
@@ -172,19 +178,21 @@ export class Repository<
   public async remove(id: string): Promise<T | void> {
     if (this.isInit) {
       try {
-        this.startLoad();
+        if (this.service.ACL.remove.includes(this.service.group)) {
+          this.startLoad();
 
-        const item: T | void = await this.service.remove(id);
+          const item: T | void = await this.service.remove(id);
 
-        if (item) {
-          runInAction(`[ ${this.constructor.name} ][ REMOVE ][ SUCCESS ]`, () => {
-            this.collection.delete(item.id);
-          });
-        } else {
-          throw new Error(`item: ${Object.prototype.toString.call(item)}`);
+          if (item) {
+            runInAction(`[ ${this.constructor.name} ][ REMOVE ][ SUCCESS ]`, () => {
+              this.collection.delete(item.id);
+            });
+          } else {
+            throw new Error(`item: ${Object.prototype.toString.call(item)}`);
+          }
+
+          return item;
         }
-
-        return item;
       } catch (error) {
         console.error(`[ ${this.constructor.name} ][ UPDATE ][ ERROR_MESSAGE: ${error.message || error} ]`);
